Add support for escaped identifiers
This commit mirrors https://github.com/segmentio/fql/pull/53

One slight difference is that `fql` stores the escaped representation of
the identifier internally but this implementation stores the raw
representation of the identifier. `fql` behaves differently because it
proxies the identifier to the `gjson` library during evaluation, which
needs the escaped version, whereas `fql-ts` proxies to and from the raw
UI, where users will not expect to see the escaped version of an
identifier.